+++ ruby/CHANGELOG.md
@@ -4,6 +4,7 @@ This is the changelog for the `essence` gem. For now, most of the updates will b
 
 ### Unreleased
 
+- feature: Adding `Checkbox` component
 - feature: Adding `Switch` component
 - feature: Adding `Separator` component